real_estate_bot_clean/requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
setuptools>=68.0.0
wheel>=0.40.0

# API and Networking
aiohttp>=3.8.0
httpx>=0.24.0
requests>=2.28.0
grpcio>=1.54.0
protobuf>=4.22.0

# Data Processing
python-dateutil>=2.8.2
pytz>=2023.3
statsmodels>=0.14.0
scipy>=1.11.0

# Storage and Caching
diskcache>=5.6.1
aiofiles>=23.1.0
redis>=4.5.0
pymongo>=4.3.0

# Google Integration
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.95.0

# Security
cryptography>=40.0.0
pyjwt>=2.6.0
python-jose>=3.3.0
passlib>=1.7.4
circuitbreaker>=1.4.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
structlog>=23.1.0
typing-extensions>=4.5.0
asyncio>=3.4.3

# Testing and Development
pytest>=7.0.0
pytest-asyncio>=0.23.5
pytest-cov>=4.1.0
pytest-mock>=3.10.0
pytest-benchmark>=4.0.0
responses>=0.23.0

# Code Quality
black>=23.7.0
flake8>=6.1.0
isort>=5.12.0
mypy>=1.3.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.2.0
